requirements.txt
aadict==0.2.3
alembic==1.4.2
asset==0.6.13
Babel==2.8.0
bcrypt==3.1.7
beautifulsoup4==4.9.0
bidict==0.19.0
bleach==3.1.5
blinker==1.4
boto3==1.13.7
botocore==1.16.7       # Required by boto3
certifi==2020.4.5.1      # Required by requests
cffi==1.14.0
chardet==3.0.4
Click==7.1.2
dictalchemy==0.1.2.7
docutils==0.15.2
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.1
Flask-SSLify==0.1.5
Flask-WTF==0.14.3
gitdb2==4.0.2
GitPython==3.1.2
globre==0.1.5
gunicorn==20.0.4
hurry.filesize==0.9
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.9.5
lxml==4.5.0
Mako==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
passlib==1.7.2
passwordmeter==0.1.8
psycopg2-binary==2.8.5
py-mini-racer==0.2.0
py-trello==0.16.0
pycparser==2.20
python-dateutil==2.8.1  # Required by alembic
python-dotenv==0.13.0
python-editor==1.0.4
python-slugify==4.0.0
pytz==2020.1            # Required by Babel
raven==6.10.0            # Client for Sentry, which does app monitoring
requests==2.23.0
s3transfer==0.3.3
six==1.14.0             # Required by aadict
smmap2==3.0.1
speaklater==1.3
SQLAlchemy==1.3.16
sqreen>=1.19.0,<2.0
Unidecode==1.1.1       # Required by python-slugify
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
password-strength>=0.0.3.post2
email-validator==1.1.0
